Reorganize drivers
